requirements.txt
PyQt5==5.15.6
bcrypt==3.2.0
pyserial==3.5
